Setting up the models for loading FICOR data
